Program is able to demonstrate adding new stock and user can select a stock they already have or a new one.

+++ src/com/atden04/java/share_database_app/models/Dividend.java
@@ -6,14 +6,20 @@ import javafx.beans.property.SimpleStringProperty;
 
 public class Dividend {
     private SimpleStringProperty date;
-    private SimpleIntegerProperty quantity;
+    private SimpleFloatProperty quantity;
 
     private SimpleFloatProperty paymentRate;
     private SimpleFloatProperty value;
 
-    public Dividend(String newDate, int newQuantity, float newCost){
+    /**
+     * blah blah
+     * @param newDate Date Dividend occurred
+     * @param newQuantity quantity
+     * @param newCost cost of share.
+     */
+    public Dividend(String newDate, float newQuantity, float newCost){
         this.date = new SimpleStringProperty(newDate);
-        this.quantity = new SimpleIntegerProperty(newQuantity);
+        this.quantity = new SimpleFloatProperty(newQuantity);
         this.value = new SimpleFloatProperty(newCost);
         this.paymentRate = new SimpleFloatProperty(this.value.get()/this.quantity.get());
     }
